MPS br/vmem tract.c: (minor cleanups):
tractSearchInChunk(): If BTGet(chunk->allocTable, i) is true, then 
necessarily so is PageIsAllocated(&chunk->pageTable[i]).  Remove this 
redundant clause from the conditional, for clarity.  (It seems to have 
got in there when the function was created by editing segSearchInChunk).
Fuller AVERs, consistently AVER_CRITICAL.

rename "ChunkOfNextAddr" as "NextChunkAboveAddr".  The old name was 
misleading: the word "Next" applies to chunks, not to addrs. 

Copied from Perforce
 Change: 171419
 ServerID: perforce.ravenbrook.com