New segment class MarkSweepSeg for leaf-only mark-and-sweep segments.
(This is a renamed and generalized version of LOSeg.)

Segments are now responsible for all their pool generation accounting:
in particular, accounting for freeing the segment is handled via
SegFinish, and no longer via the pool. This avoid duplicated logic
between the segment reclaim method and the pool finish method.