Prevent races between threads in amcssth.
The races arise when multiple threads modify exactRoots, or an object
pointed to by exactRoots, without synchronizing. Prevent races by
serializing access to exactRoots, and the objects pointed to by
exactRoots, with a mutex. This doesn't give us the maximum possible
parallelism but it is easy to see that it is correct.

Add mutexes to the test threads interface.