Workaround Apple Silicon restrictions on PROT_EXEC
Calling mmap(PROT_WRITE | PROT_EXEC) will fail due to
restrictions on marking memory executable & writable at the same time.

See apple developer documentation on this new restriction:
https://developer.apple.com/documentation/apple-silicon/porting-just-in-time-compilers-to-apple-silicon

This workaround simply stops requesting PROT_EXEC entirely on Apple
Silicon. This is not a complete solution, because it does not allow clients
to use seamlessly garbage collection to manage their executable memory.

See issue #75 (which is not fully resolved by this)

Note that any "proper" solution would still require signficant work to interact with
the new Apple APIs for JIT.
This means that it is almost certinally going to require some effort on
the part of the client, and possibly more APIs on the part of the MPS.

This commit does manage to get basic support for the Apple Silicon chips working,
and it works fine as long as you don't use the MPS to allocate
executable memory (it's fine if you do it seperately).
The example scheme interpreter builds and runs correctly on my Apple M1 :)