WebAssembly: Use unix locks when threads enabled.
When compiling with `-s USE_PTHREADS` in emscripten, Posix threads
and locks are available (assuming the WebAssembly VM supports this
extension).

This can be detected at compile time by checking for the preprocessor
definition `__EMSCRIPTEN_PTHREADS__`.

In the absence of this definition, we request a single threaded
configuration which falls back to the implementation in `lockan.c`.

This works when building `mps.c`, but does not work when
using the older Makefile-based build system as when building
`lockix.c` as a standalone file, `mpstd.h` is not included prior
to the relevant portion of `config.h` being processed.