Initial take on CI for emscripten.
This uses GitHub actions and installs the Emscripten SDK's
latest version and activates it. At some point, this should
probably change to a fixed version that is periodically
incremented once some baseline support level is determined.

This is just compiling and running what's needed for `amsss`
for now until other pieces are in place for building and
running more of everything.